requirements.txt

django>=4.2
mysqlclient
djangorestframework
djangorestframework-simplejwt
django-cors-headers
pillow
